Menu added

+++ Include/MenuFunctions.hpp
@@ -9,7 +9,7 @@ namespace MenuFunctions{
     void ClearTerminal();
     void WaitForAction();
     std::string insertPassword();
-    size_t insertNumber(const size_t& zakresL, const size_t& zakresP);
+    int insertNumber(const int& zakresL, const int& zakresP);
     std::string SetTextColor(const Color& color, const std::string& text);
     std::string SetBackgroundColor(const Color& color, const std::string& text);
 }
