+++ Src/MenuFunctions.cpp
@@ -41,8 +41,8 @@ std::string MenuFunctions::insertPassword(){
         return password;
     }
 
-size_t  MenuFunctions::insertNumber(const size_t& zakresL, const size_t& zakresP){
-    size_t choice;
+int  MenuFunctions::insertNumber(const int& zakresL, const int& zakresP){
+    int choice;
     std::cout<<"Insert number in range: "<<zakresL <<  " to " <<zakresP<<"\n>";
     while(!(std::cin>>choice) || choice<zakresL || choice >zakresP){
         std::cout<<"Please enter number only in range: "<<zakresL <<  " to " <<zakresP<<"\n>";
